fix: resolve test and build issues
