+++ go-bridge/cmd/mandala-gpt/main.go
